+++ src/AppRoutes.tsx
@@ -13,8 +13,8 @@ const Dashboard = lazy(() => import('./pages/dashboard/Dashboard'));
 const BasicDataPage = lazy(() => import('./pages/basic_data/BasicDataPage'));
 const EditorPage = lazy(() => import('./pages/editor/InstitutionPage'));
 const EditRegistration = lazy(() => import('./pages/registration/new_registration/EditRegistration'));
-const CreateProject = lazy(() => import('./pages/project/new_project/CreateProject'));
-const EditProject = lazy(() => import('./pages/project/new_project/EditProject'));
+const CreateProject = lazy(() => import('./pages/project/project_wizard/CreateProject'));
+const EditProject = lazy(() => import('./pages/project/project_wizard/EditProject'));
 const PublicRegistration = lazy(() => import('./pages/public_registration/PublicRegistration'));
 const PrivacyPolicy = lazy(() => import('./pages/infopages/PrivacyPolicy'));
 const ProjectsPage = lazy(() => import('./pages/projects/ProjectsPage'));
