+++ src/pages/basic_data/app_admin/central_import/CentralImportCandidateForm.tsx
@@ -16,10 +16,10 @@ import { BackgroundDiv } from '../../../../components/styled/Wrappers';
 import { setNotification } from '../../../../redux/notificationSlice';
 import { ImportCandidate } from '../../../../types/importCandidate.types';
 import { RegistrationTab } from '../../../../types/registration.types';
-import { getTouchedTabFields } from '../../../../utils/formik-helpers';
+import { getTouchedTabFields } from '../../../../utils/formik-helpers/formik-helpers';
 import { getTitleString } from '../../../../utils/registration-helpers';
 import { createUppy } from '../../../../utils/uppy/uppy-config';
-import { IdentifierParams, getImportCandidatePath } from '../../../../utils/urlPaths';
+import { getImportCandidatePath, IdentifierParams } from '../../../../utils/urlPaths';
 import { registrationValidationSchema } from '../../../../utils/validation/registration/registrationValidation';
 import { ContributorsPanel } from '../../../registration/ContributorsPanel';
 import { DescriptionPanel } from '../../../registration/DescriptionPanel';
