+++ src/pages/public_registration/action_accordions/PublishingAccordion.tsx
@@ -38,7 +38,7 @@ import { Registration, RegistrationStatus, RegistrationTab } from '../../../type
 import { isErrorStatus, isSuccessStatus } from '../../../utils/constants';
 import { dataTestId } from '../../../utils/dataTestIds';
 import { toDateString } from '../../../utils/date-helpers';
-import { getFirstErrorTab, getTabErrors, validateRegistrationForm } from '../../../utils/formik-helpers';
+import { getFirstErrorTab, getTabErrors, validateRegistrationForm } from '../../../utils/formik-helpers/formik-helpers';
 import { userCanPublishRegistration } from '../../../utils/registration-helpers';
 import { getRegistrationLandingPagePath, getRegistrationWizardPath, UrlPathTemplate } from '../../../utils/urlPaths';
 import { TicketMessageList } from '../../messages/components/MessageList';
